% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{add_activation}
\alias{add_activation}
\title{Add Activation Layer}
\usage{
add_activation(
  model,
  activ_fn = c("linear", "relu", "sigmoid", "softmax_crossentropy")
)
}
\arguments{
\item{model}{The model object to which this layer will be added.}

\item{activ_fn}{The name of the activation function which will be applied to
the output of the previous dense layer. One of "linear", "relu", "sigmoid",
"softmax_crossentropy". If a softmax activation function is to be used
with the categorical crossentropy loss function, use softmax_crossentropy.}
}
\description{
Add an activation layer to the model. This layer must directly follow a dense
layer, and must be directly followed by either another dense layer, or the loss
function.
}
