% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{add_dense}
\alias{add_dense}
\title{Add Dense Layer}
\usage{
add_dense(
  model,
  n_inputs,
  n_neurons,
  dropout_rate = 0,
  weight_L1 = 0,
  weight_L2 = 0,
  bias_L1 = 0,
  bias_L2 = 0
)
}
\arguments{
\item{model}{The model object to which this layer will be added.}

\item{n_inputs}{The number of inputs this layer will receive. If the first layer,
this should be equal to the number of samples fed into the neural net. If
a subsequent hidden layer, this should be equal to the number of neurons set
in the last dense layer.}

\item{n_neurons}{The desire number of neurons or nodes for the layer to contain.}

\item{dropout_rate}{The rate of dropout. Corresponds to the proportion of neurons
in the layer that will be randomly disabled.}

\item{weight_L1}{L1 regularization for the weights in this layer.}

\item{weight_L2}{L2 regularization for the weights in this layer.}

\item{bias_L1}{L1 regularization for the biases in this layer.}

\item{bias_L2}{L2 regularization for the biases in this layer.}
}
\description{
Add a dense layer to the model. Specify the correct number of inputs and the
desired number of neurons, and additional parameters as desired.
}
