% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{add_loss}
\alias{add_loss}
\title{Add Loss Layer}
\usage{
add_loss(
  model,
  loss_fn = c("mse", "mae", "binary_crossentropy", "categorical_crossentropy")
)
}
\arguments{
\item{model}{The model object to which this layer will be added.}

\item{loss_fn}{The name of the loss function which will be used as the objective
function in the training of the neural network. One of "mse" for Mean Squared Error,
"mae" for Mean Absolute Error, "binary_crossentropy", or "categorical_crossentropy".}
}
\description{
Add a loss function to the model. This should be the last layer specified in
the sequential list of layers built on top of \code{nnfR::initialize_model()}. The
previous layer must be an activation layer (\code{nnfR::add_activation()}).
}
