% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model.R
\name{print_model}
\alias{print_model}
\title{Print model information}
\usage{
print_model(model)
}
\arguments{
\item{model}{A model built on top of nnfR::initialize_model(), which is a list
of layer objects. Not a trained/fitted model.}
}
\description{
After specifying the structure of a model, print it in a nice format to
visually inspect.
}
