% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data.R
\name{sim_spiral_data}
\alias{sim_spiral_data}
\title{Simulate Spiral Data}
\usage{
sim_spiral_data(N = 100, D = 2, K = 3, random_order = FALSE)
}
\arguments{
\item{N}{the number of data points per class}

\item{K}{the number of classes}
}
\description{
Simulate spiral data for multi-class classification tasks.
Source: https://cs231n.github.io/neural-networks-case-study/
}
