% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Test Model}
\usage{
test_model(model, trained_model, X_test, y_test = NULL, metric_list = NULL)
}
\arguments{
\item{model}{A model object built ontop of the \code{nnfR::initialize_model()} function}

\item{trained_model}{The trained/fitted model object created from the
\code{nnfR::train_model()} function.}

\item{X_test}{the X inputs (the features, without the outcome variable/y/labels)}

\item{y_test}{Optional. The y outputs corresponding to the X inputs. These can be
provided if you want to generate final loss and metrics from training data. If
excluded, this function will simply make predictions from the X inputs.}

\item{metric_list}{An optional string or list of strings containing the names
of metrics to track and calculate in addition to the chosen loss function.
May include one or more of "mse", "mae", "accuracy", or "regression accuracy".}
}
\description{
Make training or testing predictions with a fitted model. This function can be
used to get predictions from a model after it is trained, or to make final
predictions on test data.
}
