% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a Neural Network Model
Fit a built neural network model onto data for a specified number of epochs.
An extensive list of function arguments allows for customization of the
training procedure.}
\usage{
train_model(
  model,
  inputs,
  y_true,
  epochs,
  optimizer = c("adam"),
  learning_rate = 0,
  lr_decay = 0,
  epsilon = 1e-07,
  ...,
  metric_list = NULL,
  validation_X = NULL,
  validation_y = NULL,
  batch_size = NULL,
  epoch_print = 100,
  batch_print = 0
)
}
\arguments{
\item{model}{A model object built ontop of the \code{nnfR::initialize_model()} function}

\item{inputs}{A matrix containing the inputs into the neural network, often
referred to as X. This should not include the labels. If not in matrix
form, consider converting with X = as.matrix(X).}

\item{y_true}{The training labels or values (y) corresponding to the inputs (X).
Can be a list or vector. For classification, the labels must be indexed
starting at 1.}

\item{epochs}{the number of training epochs}

\item{optimizer}{A string stating one of: "sgd", "adagrad", "rmsprop", or "adam".
Consider running optimizer_REMINDME() for more information on optimizers.}

\item{learning_rate}{Optimizer setting: the learning rate or alpha}

\item{lr_decay}{Optimizer setting: learning rate decay}

\item{epsilon}{Optimizer setting: epsilon, a small integer to avoid division by zero}

\item{...}{Optimizer settings: additional optimizer settings can be provided
as additional arguments. See optimizer_REMINDME() for details on which arguments
are available for which optimizers.}

\item{metric_list}{An optional string or list of strings containing the names
of metrics to track and calculate in addition to the chosen loss function.
May include one or more of "mse", "mae", "accuracy", or "regression accuracy".}

\item{validation_X}{An optional matrix containing validation data inputs (X) if desired.}

\item{validation_y}{The validation labels corresponding to validation inputs}

\item{batch_size}{An optional integer to be used if batching is desired. The
integer corresponds to the number of samples included per batch.}

\item{epoch_print}{An integer. Print epoch-wise updates to console every 'epoch_print'}

\item{batch_print}{An integer. Print batch-wise updates to console every 'batch_print'}
}
\description{
Train a Neural Network Model
Fit a built neural network model onto data for a specified number of epochs.
An extensive list of function arguments allows for customization of the
training procedure.
}
