% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classess.R
\name{classess}
\alias{classess}
\title{Classification Assessment}
\usage{
classess(truths, predictions)
}
\arguments{
\item{truths}{A vector of true labels (indexed as numbers from 1 to k)}

\item{predictions}{A vector of predicted labels (with same indexing from truths)}
}
\value{
A confusion matrix and dataframe storing classification metrics
}
\description{
This function takes as input the groud truth labels and predicted labels from
a binary or multi-class classification problem. For binary tasks, it computes
a simple confusion matrix and calculates classification metrics.
For multi-class tasks, it computes a multi-class confusion matrix and computes
the classification metrics for each class.
Current metrics are accuracy, sensitivity/recall, specificity, precision,
and F1 score.
}
